% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.Dates.R
\name{as.Dates}
\alias{as.Dates}
\title{Converting potential date to Date vector}
\usage{
as.Dates(x)
}
\arguments{
\item{x}{atomic vector}
}
\value{
recognised date vector converted to date, 
otherwise \code{x} unchanged
}
\description{
The function recognises dates in formats used by INCA and Rockan.
}
\examples{
as.Dates(c(1212121212, "20000101", "2014-10-15", 5806))
}
\seealso{
\link{as.Date}
}
