% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{documents}
\alias{documents}
\title{Download and possibly open INCA documentation}
\usage{
documents(..., dir = ".", max_open = 3)
}
\arguments{
\item{...}{Arguments passed on to \code{find_documents}
\describe{
  \item{pattern}{part of document name to look for}
  \item{x}{name of register to look for}
}}

\item{dir}{directory where to save files}

\item{max_open}{maximum number of files to open automatically 
(only on Mac OS X). Set to 0 to avoid any opening of files.}
}
\value{
Nothing. The function is called for its side effects.
}
\description{
Download and possibly open INCA documentation
}
\examples{
\dontrun{
documents("lunga", "uppfoljning")
}
}
