% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{dplyr_methods}
\alias{dplyr_methods}
\alias{mutate_.incadata}
\alias{arrange_.incadata}
\alias{filter_.incadata}
\alias{rename_.incadata}
\alias{select_.incadata}
\alias{slice_.incadata}
\alias{summarise_.incadata}
\alias{summarize_.incadata}
\alias{group_by_.incadata}
\title{dplyr methods for INCA data}
\usage{
mutate_.incadata(.data, ...)

arrange_.incadata(.data, ...)

filter_.incadata(.data, ...)

rename_.incadata(.data, ...)

select_.incadata(.data, ...)

slice_.incadata(.data, ...)

summarise_.incadata(.data, ...)

summarize_.incadata(.data, ...)

group_by_.incadata(.data, ...)
}
\arguments{
\item{.data, ...}{arguments passed to original dplyr-methods}
}
\description{
Verbs from the dplyr package can be used for incadata directly, but 
the incadata object will then loose its class. 
These methods will preserve the class.
}
\keyword{internal}
