% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_ex_data.R
\docType{data}
\name{ex_data}
\alias{ex_data}
\title{Example data from INCA}
\format{A data frame with 497 rows and 433 variables}
\usage{
ex_data
}
\description{
Note that all data has been anonymised! There is no logical relation 
between variables, hence \code{x_Beskrivning} and \code{x_Varde} 
does not match.
}
\keyword{datasets}
