% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_documents.R
\name{find_documents}
\alias{find_documents}
\title{List URLS to documents for a register}
\usage{
find_documents(url, doc = NULL)
}
\arguments{
\item{url}{url to web page where to look for documents}

\item{doc}{(part of) document name to look for}
}
\value{
names character vector with urls to documents
}
\description{
List URLS to documents for a register
}
\examples{
\dontrun{
find_documents(find_register("all"))
find_documents(find_register("peniscancer"), "uppfoljning")
}
}
