% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_links.R
\name{find_links}
\alias{find_links}
\title{Find links from web page}
\usage{
find_links(url, select = NULL)
}
\arguments{
\item{url}{URL to web page with links (must be under 'www.cancercentrum.se')}

\item{select}{select only links matching specified pattern}
}
\value{
Named character vector with absolute URLs to links found 
on 'www.cancercentrum.se'
}
\description{
Find links from web page
}
\examples{
\dontrun{
# Find e-mailadresses to spam
 find_links(
   "https://cancercentrum.se/vast/om-oss/kontakta-oss/", 
   "mailto:"
 )
}
}
