% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_register.R
\name{find_register}
\alias{find_register}
\title{Find register by name}
\usage{
find_register(reg = NULL)
}
\arguments{
\item{reg}{name of register to look for}
}
\value{
Named character vector with URL to specified register
}
\description{
The specified name does not need to be exact since a search algorithm is
applied to match existing registers. Names of the registers
}
\examples{
find_register("all")
\dontrun{
find_register("kronisk") # More than one possible alternative
}
}
