% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{id}
\alias{id}
\title{Add id variables to data frame}
\usage{
id(x, id = c("persnr", "pnr", "pat_id", "pn", "id"), ignore.case = TRUE)
}
\arguments{
\item{x}{data frame}

\item{id}{names of a possible id variable found in \code{x}}

\item{ignore.case}{should name matching be done regardless of character case?}
}
\value{
Character variable with either the first name from \code{id} found in 
\code{x} or \code{rownames(x)} if no named column found.
}
\description{
Construct id variable for patient data.
}
