% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.inca.R
\name{is.inca}
\alias{is.inca}
\title{Check if R is running from INCA}
\usage{
is.inca(logical = TRUE)
}
\arguments{
\item{logical}{Should the return value be a simple boolean whether we
are running from INCA or not?}
}
\value{
Either \code{TRUE}\code{FALSE} if \code{logical = TRUE} or one of
"PROD", "TEST" or "LOCAL" depending on were R is running
(if \code{logical = FALSE})
}
\description{
Check if R is running from INCA
}
\examples{
is.inca()
}
