% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_leadtime.R
\name{lt}
\alias{lt}
\title{Lead time from one date to another}
\usage{
lt(from, to, neg.na = TRUE)
}
\arguments{
\item{from, to}{start and stop dates (in formats that can be coerced by
\code{\link{as.Dates}}).}

\item{neg.na}{should negative lead times be set to \code{NA}?}
}
\value{
Numeric vector
}
\description{
Lead time from one date to another
}
\examples{
lt(from = Sys.Date(), to = Sys.Date() + 10)
}
