% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_ex_data.R
\docType{data}
\name{ex_data}
\alias{ex_data}
\title{Synthetic example data from INCA}
\format{
A data frame (not an object of class \code{incadata} 
with 497 rows and 433
  variables
}
\usage{
ex_data
}
\description{
A data set resembling the typical form of INCA data. Variable names are real
but all data has been carefully anonymized!
}
\details{
All data is random! There is no logical relation between any
variables, not even between \code{x_Beskrivning} and \code{x_Varde}!
}
\examples{
as.incadata(ex_data)
}
\keyword{datasets}
