% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_documents.R
\name{find_documents}
\alias{find_documents}
\title{List all documents for a register}
\usage{
find_documents(x, pattern = NULL)
}
\arguments{
\item{x}{name of register to look for}

\item{pattern}{part of document name to look for}
}
\value{
data frame with names of documents and corresponding URL:s for 
specified register
}
\description{
List all documents for a register
}
\examples{
\dontrun{
find_documents("all")
find_documents("peniscancer", "uppfoljning")
}
}
