% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_register.R
\name{find_register}
\alias{find_register}
\title{Find register by name}
\usage{
find_register(x = NULL)
}
\arguments{
\item{x}{name of register to look for}
}
\value{
Name of existing register according to <www.incanet.se>
}
\description{
Find register by name
}
\examples{
find_register("all") # "akut lymfatiskt leukemi all"
\dontrun{
find_register("kronisk") # More than one possible alternative
}
}
