% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_leadtime.R
\name{lt}
\alias{lt}
\title{Lead time from one date to another}
\usage{
lt(from, to, neg = FALSE)
}
\arguments{
\item{from, to}{start and stop dates (in formats that can be recognized as RCC dates).}

\item{neg}{except negative lead times (set to \code{NA} if \code{neg = FALSE})?}
}
\value{
Numeric vector
}
\description{
Lead time from one date to another
}
\examples{
lt("2017-02-10", "2017-02-16") # 6
lt("2017-02-16", "2017-02-10") # negative lead times ignored by default
lt("2017-02-16", "2017-02-10", TRUE) # -6
}
